src/api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
langchain_openai
langchain_community
flask
azure-search-documents
azure.identity
python-dotenv
azurefunctions.extensions.bindings.blob